metaphor-api/requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Flask==2.3.3
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
jellyfish==1.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
metaphor-python==0.1.16
networkx==3.1
numpy==1.25.2
python-dotenv==1.0.0
regex==2023.8.8
requests==2.31.0
segtok==1.5.11
tabulate==0.9.0
urllib3==2.0.4
Werkzeug==2.3.7
yake @ git+https://github.com/LIAAD/yake@374fc1c1c19eb080d5b6115cbb8d4a4324392e54
zipp==3.16.2
